spec-manager/requirements.txt
requests
pyyaml
deepdiff
jinja2